fixes made as suggested in code review - restful routes, refactoring for efficiency
