+++ blueprints/add_drink.py
@@ -3,13 +3,13 @@ from datetime import datetime
 from flask import Blueprint, session, render_template
 from bson import ObjectId
 
-from .utils import *
+import utils
 from .api import get_ingredients_by_type, get_flavors
 
 add_drink = Blueprint('add_drink', __name__)
 
 
-@add_drink.route('/add_cocktail/', methods=["GET", "POST"])
+@add_drink.route('/cocktails/new', methods=["GET", "POST"])
 def new_drink():
     """
     render the form for adding a new cocktail to the database
@@ -21,26 +21,25 @@ def new_drink():
         return redirect(url_for("index"))
 
 
-@add_drink.route('/cocktail_processing/', methods=["POST"])
+@add_drink.route('/cocktails/', methods=["POST"])
 def add_new_drink_to_db():
     """
     function called from the cocktail form to add a drink to the
     database using AJAX
     """
-    ingredients = json.loads(get_ingredients_by_type(None))
-    flavors = json.loads(get_flavors())
+
     data = request.data
     dataDict = json.loads(data)
     # function returns an array
     # index 0: list of flavors
     # index 1: is a list of ingredients
-    ingredients_and_flavors = get_ingredient_and_flavor_list(dataDict)
+    ingredients_and_flavors = utils.get_ingredient_and_flavor_list(dataDict)
     connection = mongo_connect()
     connection["cocktails"].insert_one({
         "name": dataDict["name"],
         "description": dataDict["description"],
-        "flavor_tags": flavorIds,
-        "ingredients": ingredientIds,
+        "flavor_tags": ingredients_and_flavors[0],
+        "ingredients": ingredients_and_flavors[1],
         "method": dataDict["instructions"],
         "glass": dataDict["glass"],
         "equipment": dataDict["equipment"],
@@ -54,34 +53,8 @@ def add_new_drink_to_db():
         "preferred_spirits": [],
         "image_url": dataDict["image_url"]
     })
-    return "success"
+    resp = jsonify(success=True)
+    return resp
 
 # Helper Funcitons to add new flavors and ingredients
 # and return there ID's from the DB
-
-
-def add_ingredient_return_id(name, type):
-    connection = mongo_connect()
-    ingredients = connection["ingredients"]
-    ingredients.insert_one(
-        {
-            "name": name,
-            "type": type,
-        })
-    newIngredient = ingredients.find_one({
-        "name": name
-    })
-    return newIngredient["_id"]
-
-
-def add_flavor_return_id(name):
-    connection = mongo_connect()
-    flavors = connection["flavors"]
-    flavors.insert_one(
-        {
-            "name": name,
-        })
-    new_flavor = flavors.find_one({
-        "name": name
-    })
-    return new_flavor["_id"]
